+++ src/main/java/ai/aecode/testweb/entities/UserResult.java
@@ -25,15 +25,15 @@ public class UserResult {
     @Column(name = "result_manager")
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "0.0")
     @JsonSerialize(using = ToStringSerializer.class)
-    private Double result_manager;
+    private Integer result_manager;
     @Column(name = "result_developer")
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "0.0")
     @JsonSerialize(using = ToStringSerializer.class)
-    private Double result_developer;
+    private Integer result_developer;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "0.0")
     @JsonSerialize(using = ToStringSerializer.class)
     @Column(name = "result_executor")
-    private Double result_executor;
+    private Integer result_executor;
     @Column(name = "person_type_description")
     private String person_type_description;
 
@@ -47,7 +47,7 @@ public class UserResult {
     public UserResult() {
     }
 
-    public UserResult(int id_userresult, UserProfile userProfile, PersonType personType, Double result_manager, Double result_developer, Double result_executor, String person_type_description, Set<Skill> skill) {
+    public UserResult(int id_userresult, UserProfile userProfile, PersonType personType, Integer result_manager, Integer result_developer, Integer result_executor, String person_type_description, Set<Skill> skill) {
         this.id_userresult = id_userresult;
         this.userProfile = userProfile;
         this.personType = personType;
@@ -82,27 +82,27 @@ public class UserResult {
         this.personType = personType;
     }
 
-    public Double getResult_manager() {
+    public Integer getResult_manager() {
         return result_manager;
     }
 
-    public void setResult_manager(Double result_manager) {
+    public void setResult_manager(Integer result_manager) {
         this.result_manager = result_manager;
     }
 
-    public Double getResult_developer() {
+    public Integer getResult_developer() {
         return result_developer;
     }
 
-    public void setResult_developer(Double result_developer) {
+    public void setResult_developer(Integer result_developer) {
         this.result_developer = result_developer;
     }
 
-    public Double getResult_executor() {
+    public Integer getResult_executor() {
         return result_executor;
     }
 
-    public void setResult_executor(Double result_executor) {
+    public void setResult_executor(Integer result_executor) {
         this.result_executor = result_executor;
     }
